Add WatcherGroup class to _watcher module
