+++ capsula/_watcher/__init__.py
@@ -1,4 +1,4 @@
-__all__ = ["TimeWatcher", "UncaughtExceptionWatcher", "WatcherBase"]
-from ._base import WatcherBase
+__all__ = ["TimeWatcher", "UncaughtExceptionWatcher", "WatcherBase", "WatcherGroup"]
+from ._base import WatcherBase, WatcherGroup
 from ._exception import UncaughtExceptionWatcher
 from ._time import TimeWatcher
